feat: add getTimeNow to component and post reply
