+++ src/apis/apis.ts
@@ -1,6 +1,7 @@
 import axios from 'axios';
 import { FormDataType as CreateContractDataType } from '../components/landLordManagement/offcanvases/CreateContract';
 import { ReviewPostDataType } from '../components/reviews/MyReview';
+import { ReplyDataType } from '../components/reviews/ReplyReview';
 
 const getToken = () => {
   const token = localStorage.getItem("authToken");
@@ -134,4 +135,4 @@ export const apiUserInfoGet = () => userRequest.get('/tenant/info',getToken()); 
 // 評價-共用相關的 api
 export const apiCommentList = (querystring: string) => commentRequest.get(`/common/list/all?${querystring}`, getToken()); // ACC-1
 export const apiCommentPost = (data: ReviewPostDataType, orderId: number) => commentRequest.post(`/common/${orderId}`, data, getToken()); // ACC-5
-export const apiCommentReply = (data: any, commentId: number) => commentRequest.post(`/common/reply/${commentId}`, data, getToken()); // ACC-6
+export const apiCommentReply = (data: ReplyDataType, commentId: number) => commentRequest.post(`/common/reply/${commentId}`, data, getToken()); // ACC-6
